Separate module sort logic to ModuleGraph
